requirements.txt
moviepy
pydub
numpy
Pillow
yt-dlp
syncedlyrics
librosa
soundfile
